+++ test/Sentry.Tests/Internals/Http/CachingTransportTests.cs
@@ -192,7 +192,7 @@ namespace Sentry.Tests.Internals.Http
 
             innerTransport
                 .SendEnvelopeAsync(Arg.Any<Envelope>(), Arg.Any<CancellationToken>())
-                .Returns(_ =>  new ValueTask(Task.FromException(exception)));
+                .Returns(_ =>  Task.FromException(exception));
 
             await using var transport = new CachingTransport(innerTransport, options);